requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
Pillow
google-generativeai
pydantic-settings
fuzzywuzzy
python-Levenshtein